+++ frontend/app/user/components/UserStatistics.tsx
@@ -4,7 +4,7 @@ import { cn } from "@/lib/utils";
 import Link from "next/link";
 import prettyBytes from "pretty-bytes";
 import { HTMLAttributes } from "react";
-import { UserStats } from "../types";
+import { UserStats } from "../../../lib/types/User";
 import { BrainConsumption } from "./BrainConsumption";
 import { DateComponent } from "./Date";
 import BrainSpaceChart from "./Graphs/BrainSpaceChart";
