+++ frontend/lib/types/Document.ts
