+++ frontend/lib/types/Message.ts
